src/services/table-service/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
flask-openapi3==3.1.3
httplib2==0.22.0
pydantic==2.8.2
gunicorn==22.0.0
PyJWT==2.9.0
numpy==2.0.1
pymongo==4.10.1
python-dotenv==1.0.1